refactor: promote simulation timeline workbench API
