+++ robot_sf/analysis_workbench/__init__.py
@@ -9,6 +9,15 @@ from robot_sf.analysis_workbench.real_trace_validation_contract import (
     check_real_trace_validation_contract,
     load_real_trace_validation_contract,
 )
+from robot_sf.analysis_workbench.simulation_timeline import (
+    SIMULATION_TIMELINE_SCHEMA_FILE,
+    SIMULATION_TIMELINE_SCHEMA_VERSION,
+    SimulationTimelineValidationError,
+    build_simulation_timeline,
+    load_simulation_timeline_schema,
+    validate_simulation_timeline,
+    write_simulation_timeline,
+)
 from robot_sf.analysis_workbench.simulation_trace_export import (
     SIMULATION_TRACE_EXPORT_SCHEMA_VERSION,
     SimulationTraceExport,
@@ -36,12 +45,15 @@ from robot_sf.analysis_workbench.trace_failure_predicates import (
 __all__ = [
     "CONTRACT_EVIDENCE_BOUNDARY",
     "REAL_TRACE_VALIDATION_CONTRACT_SCHEMA_VERSION",
+    "SIMULATION_TIMELINE_SCHEMA_FILE",
+    "SIMULATION_TIMELINE_SCHEMA_VERSION",
     "SIMULATION_TRACE_EXPORT_SCHEMA_VERSION",
     "TRACE_ANNOTATION_SET_SCHEMA_VERSION",
     "TRACE_FAILURE_PREDICATE_SCHEMA_VERSION",
     "PredicateCompatibility",
     "RealTraceValidationContractError",
     "RealTraceValidationContractReport",
+    "SimulationTimelineValidationError",
     "SimulationTraceExport",
     "SimulationTraceExportValidationError",
     "TraceAnnotationSet",
@@ -49,13 +61,17 @@ __all__ = [
     "TraceFailurePredicate",
     "TraceFailurePredicateDefinition",
     "aggregate_trace_failure_predicate_tables",
+    "build_simulation_timeline",
     "build_trace_failure_predicate_definitions",
     "check_real_trace_validation_contract",
     "extract_trace_failure_predicates",
     "load_real_trace_validation_contract",
+    "load_simulation_timeline_schema",
     "load_simulation_trace_export",
     "load_trace_annotation_set",
     "render_trace_failure_predicate_markdown",
     "simulation_trace_export_from_dict",
     "trace_annotation_set_from_dict",
+    "validate_simulation_timeline",
+    "write_simulation_timeline",
 ]
